# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Magpie Interactive Reading Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  message: string;
  code?: string;
}

export interface UserProfile {
  id: string;
  email: string;
  displayName: string;
  avatarUrl?: string | null;
  createdAt: string;
}

export interface UpdateProfileRequest {
  displayName?: string;
  avatarUrl?: string | null;
}

export interface Story {
  id: string;
  title: string;
  genre: string;
  description: string;
  longDescription: string;
  coverGradient: string;
  tags: string[];
  rating: string;
  chapterCount: number;
  readingTime: string;
  featured: boolean;
  rank?: number | null;
  initialPrompt: string;
}

export type StorySessionStatus =
  (typeof StorySessionStatus)[keyof typeof StorySessionStatus];

export const StorySessionStatus = {
  active: "active",
  completed: "completed",
  abandoned: "abandoned",
} as const;

export interface StorySession {
  id: string;
  storyId: string;
  userId: string;
  status: StorySessionStatus;
  currentNodeId?: string | null;
  nodeCount: number;
  createdAt: string;
  updatedAt: string;
  story: Story;
}

export interface Choice {
  id: string;
  text: string;
  consequence?: string;
}

export interface StoryNode {
  id: string;
  sessionId: string;
  parentNodeId?: string | null;
  choiceMade?: string | null;
  narrativeText: string;
  choices: Choice[];
  nodeIndex: number;
  createdAt: string;
}

export interface SessionWithNodes {
  session: StorySession;
  nodes: StoryNode[];
  currentNode?: StoryNode | null;
}

export interface CreateSessionRequest {
  storyId: string;
}

export interface ContinueSessionRequest {
  choiceId: string;
  choiceText: string;
}

export interface GenerateSegmentRequest {
  storyTitle: string;
  genre: string;
  previousContext: string;
  choiceMade: string;
  nodeIndex: number;
}

export interface GeneratedSegment {
  narrativeText: string;
  choices: Choice[];
  isEnding: boolean;
}

export type ListStories200 = {
  stories: Story[];
};

export type ListSessions200 = {
  sessions: StorySession[];
};