# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Magpie Interactive Reading Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/profile:
    get:
      operationId: getProfile
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
      required: [message]
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, email, displayName, createdAt]
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
...