# lib/api-zod/src/generated/types/updateProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Magpie Interactive Reading Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProfileRequest {
  displayName?: string;
  avatarUrl?: string | null;
}